==> code/01_packagemanagment.R <==


install.packages("here", version='1.0.1')
install.packages("tidyverse", version='2.0.0')
install.packages("fixest", version='0.11.1')
install.packages("huxtable", version='5.5.2')
install.packages("flextable", version='0.8.6')
